requirements.txt
# Core ML Dependencies
torch==2.1.0
transformers==4.36.0
datasets==2.14.0
tokenizers==0.15.0
accelerate==0.24.1
peft==0.7.1
bitsandbytes==0.41.3

# Fine-tuning and Optimization
trl==0.7.4
optimum==1.14.0
auto-gptq==0.5.1

# API and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2

# Data Processing
pandas==2.1.3
numpy==1.24.4
pyarrow==14.0.1
scikit-learn==1.3.2

# Evaluation Metrics  
evaluate==0.4.1
rouge-score==0.1.2
sacrebleu==2.3.1
bert-score==0.3.13
sentence-transformers==2.2.2

# Monitoring and Logging
wandb==0.16.0
tensorboard==2.15.1
prometheus-client==0.19.0
structlog==23.2.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
hydra-core==1.3.2

# Caching and Storage
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Jupyter and Analysis
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# CUDA and GPU (optional)
# flash-attn==2.3.6  # Uncomment for GPU optimization
# deepspeed==0.12.3  # Uncomment for distributed training

# Production
gunicorn==21.2.0
psutil==5.9.6
python-multipart==0.0.6

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Optional: Cloud providers
# boto3==1.34.0      # AWS
# google-cloud-storage==2.10.0  # GCP
# azure-storage-blob==12.19.0   # Azure